Improve frontend folder structure Move IconProvider to components Rectify Wrong Imports
